Regenerate DBIC schema
When I first started working on PearlBee, I was unsure whether the best
approach would be to manage the DB schema outside DBIC (and use DBIC
Schema Loader to generate the classes), or to only use DBIC, and have it
deploy the SQL databases for me. I tried picking the latter, and deleted
all the SQL files. But at this point, I miss several things from the
first approach (in no particular order):

1. Transparency - It's easy to know what the database is doing when you
tell it what to do.

2. Familiarity - I am more comfortable writing SQL to define my schema,
even if querying the data itself is generally much easier in DBIC.

3. Versioning - Sqitch makes database versioning very easy.

4. Consistency - DBIC Schema Loader always uses the same style to dump
the code, generates the same pod, and so on. Even if we run perltidy
everywhere, certain things may remain different (e.g. qw() vs qw//).

So I created this gen-dbic-schema script, which uses
DBIx::Class::Schema::Loader to regenerate the schema. I had to run it
with the 'overwrite local changes' flag on, because the checksums
mismatched. Mostly because of spacing changes (perltidy), but also some
other modifications. I tried to keep everything working the same, and
the tests pass (not that it means much at this point :D ).