Remove category and tag classes (resultset, dashboard, etc)
My plan for categories and tags is the following:

At least on the short term, categories don't exist, only tags.

A post can have multiple tags. It should be possible to filter by tags.
Multiple posts can have the same tag. If you think about it, that
actually sounds an awful lot like categories.

One difference is that usually tags are free text, and categories are
predefined. Also, on sidebars and menus, it's common to list all the
categories in a particular blog, while with tags that's usually
impractical (since it's likely that there is a big number of tags).

One could have some sort of flag to say that a particular tag is
"sticky", or special in some way. So that those tags appear in menus,
and in the forms to create or update posts (just like categories).

But either way, I can have a perfectly fine blog with only tags for the
time being. If/when I need categories, I can probably use the same
database.

In any case, this commit includes dashboard for categories and tags,
pages to display all available tags and categories, resultsets for them
that are not used anymore.