Implement 'verify your email' process
It's still very rudimentary, but it is triggering an email to the user,
with a link to /sign-up/confirm with a token. Clicking that link
activates the user.

This is a bit different than the previous way to sign up, which required
an admin to approve the registration. I find that a bit awkward, and I
don't think our main client (blogs.perl.org) will require that for now.
Unless too many bots? But the captcha is there to stop it anyway... I
find it better to remove the bad guys manually than to require approval
for everyone.

The test passes, and cover the "happy flow" of registering, getting the
email, clicking the link, and checking that the status went from pending
to activated.

There's still a long way to go!

In the registration process, there's another awkward thing: the user
can't pick a password, but a random one is generated for them. Really
weird.

On that note, we also need to have a reset password feature.

We need to have ways to override templates. Everywhere, of course, but
it would be nice if the theme could override just one template, and
accept the default for the rest. Have a fallback system.

The fact that the admin gets an email for every registration makes me
think that we also need a notification system, where users can subscribe
to notifications to certain events (if their role allows it), and get
emails about it.