Include password in signup, remove name
When the user signs up, from now on, they will set their password right
away. This is what most registration forms do anyway. The previous
behaviour was weird... the password was autogenerated, and I'm not
really sure how the user was supposed to know it. Maybe it was supposed
to be mailed, or maybe it was supposed to be set on first login, but
either way, it was incomplete in the current implementation, and it was
confusing.

I also removed `name` from the registration, since it's not strictly
required. My idea is that we can address the user by their username on
emails and such, and if they set their `name` on their dashboard, we can
use that. The username is required to be able to set up a "blog" for
them on e.g. https://blogs.perl.org/<username>/.

I didn't fix all the oddities related to password everywhere, since I
don't fully understand it. For example, if the user is deactivated, they
can still login but must change their password. I don't know what it's
meant for, but I kept it there. Also, if a new user is created by the
admin, using the dashboard, a password is auto generated and emailed to
them. I won't change that for now.