Reset password
This patch implements "forgot password" functionality in PearlBee.

Since we already had a token system using registration_token table, I
chose to reuse that for the forgot password feature. For this, I added a
column `reason` to registration_token. Reason as in, the reason for the
token, what it will be used for. I wanted to avoid the overly generic
and meaningless `type`.

The patch includes the SQL schema change, the related DBIC change, the
Perl backend, the TT pages, the email with the reset link, and the test.
The test covers basically the "happy path", not checking for all the
invalid inputs. Maybe later it will be worth to cover more, but not
right now.