Small updates in schema
I'm not keeping compatibility at this point; I don't assume anybody is
using this live yet, since I never announced it was ready. So for now,
I'm OK with changing schema in place instead of using Sqitch features.
Obviously I will be a lot more careful in the future.

I'm dropping `cover` from posts, especially as a required field. I think
my initial thought on this was to have a default image, and that would
be why it would be required. But even with a default, it makes no sense
to require it in the database level. Also, I think it's a feature we can
add later, so I'm dropping it completely from the first iteration.

I'm adding the field `meta` to posts, where I intend to add metadata for
a post. This is a lot more flexible and performant than creating a
separate table. So plugins could add extra metadata, add it would be
possible to index those in particular installations. Sounds like an
approapriate place to have no rigid schema.

It does have a drawback that for other backends (meaning, not
PostgreSQL) it will be harder to maintain compatibility. But we could
still have it as a blob in SQLite for example, in the future.

User types also updated, just making it a little more cohesive.