Replace `status` column in user with booleans
Again, not caring for not breaking existing applications since I don't
expect anything to be live yet. So I'm just freely changing the schema.

This new format makes more sense because I wanted to introduce a new
verification step. So status `pending` would become
`pending-email-validation`, and then `pending-peer-validation`? That
would be weird. Also, if a user wants to change their email, we'll have
to revalidate it. So will it go back to pending? It shouldn't. We should
just change the flag: verified_email to false.

This enables us to be more flexible on how we want to deal with it.