Enable logged in users to update their data
This commit enables the logged in user to update her/his data. It covers
`email`, `name` and `password` fields.

Now, `password` is a funny one. I'm using PATCH according to RFC 7396.
It is nice because it's unambiguous, but it also leaves no space for
sending `current_password` for example, and much less
`confirm_password`. So, the UI would have to verify these things. I'll
consider here that if the user is logged in, they can change the email
anyway. Changing the email is enough to hijack the account anyway. Thus,
getting the current_password before changing would be good, but the
overall system wouldn't be more secure. Also, sending a confirm_password
is nice from a UI perspective, but from the API perspective, not very
valuable... So it's better to handle it in the front-end.

In essence, it's hard to find examples for handling password properly
with RESTful APIs, because it's generally not part of public APIs. This
doesn't have to be PearlBee's public API, but we have to have a way to
change it anyway, internally. I'm just choosing to do it over REST to
have a decoupled dashboard front-end.