Use strftime instead of iso8601 DateTime method
DateTime's own documentation, as of the time of this commit, states that
iso8601() doesn't exactly emit a proper ISO 8601 date, since it lacks
timezone.

Checking the format that PostgreSQL returns, it sounds like it has a
very good compromise: it implements ISO 8601 according to RFC 3339,
which allows replacing T with a space character, to increase
readability. I find that a good standard, so I'll use it for our
responses too.

For now, that means using strftime("%F %T%z"). Maybe we can move that to
a central place, but for now it's not too bad to copy paste that
snippet.