Refactor ´search_posts´ to be used in other endpoints
I'm not really happy with this subroutine search_posts, but it does what
I need for now.

The idea is to have a way to format the response for retrieving posts in
multiple endpoints: either getting a list of posts, or a single one
after creating, or searching by id. I would also like to have few
queries being sent to the database, and be as fast as possible. (Of
course, depending on how one would like to scale this later on, they
might want to break the query apart so as to not have joins -- for now,
I'd like to use the database features as well as possible, since it
makes my life easier at the moment.)

What this patch does, more specifically, is to allow search_posts to be
used for a single id, and it ensures that results are only retrieved by
users 'verified_by_peers' by default. Of course, there are endpoints
that will have access to posts even if they are not published, so there
is a scape flag for that.

What I don't like about this is that it seems like I'm building my own
ORM on top of DBIC, which is not my goal at all. So possibly this
function should be split and refactored, but I'm not exactly sure how
yet.